another fix for travis

+++ c_plus_plus/src/model/classifier/classifier.cpp
@@ -13,8 +13,8 @@ std::vector<int> Classifier::classify(std::vector<SourcePosition> audioPositions
         {
             for(size_t j = 0; j < imagePositions.size(); j++)
             {
-                if(std::abs(audioPositions[i].azimuth - imagePositions[j].azimuth) > rangeThreshold
-                        || std::abs(audioPositions[i].elevation - imagePositions[j].elevation) > rangeThreshold)
+                if(std::abs(audioPositions[i].azimuth - imagePositions[j].azimuth) > static_cast<float>(rangeThreshold)
+                        || std::abs(audioPositions[i].elevation - imagePositions[j].elevation) > static_cast<float>(rangeThreshold))
                 {
                     sourcesToSuppress.push_back(static_cast<int>(i));
                 }
